+++ tests/unit/test_timeseries_convertor.py
@@ -4,7 +4,8 @@ from unittest.mock import patch
 import numpy as np
 import pytest
 
-from scmdata.time import InsufficientDataError, TimeseriesConverter
+from scmdata.errors import InsufficientDataError
+from scmdata.time import TimeseriesConverter
 
 
 @patch("scmdata.time.has_scipy", False)
@@ -30,6 +31,11 @@ def test_short_data(combo):
         combo.extrapolation_type,
     )
     for a in [[], [0], [0, 1]]:
+        if len(a) == 1 and combo.extrapolation_type == "constant":
+            # Handle a single value and constant extrapolation differently
+            # 0 or 2 values should be handled as expected
+            continue
+
         with pytest.raises(InsufficientDataError):
             timeseriesconverter._convert(np.array(a), combo.source, combo.target)
 
@@ -109,7 +115,11 @@ def test_extrapolation_with_nans():
     np.testing.assert_allclose(c.convert_from(source_vals), target_vals, rtol=1e-3)
 
 
-def test_not_enough():
+@pytest.mark.parametrize("count", [0, 1, 2])
+@pytest.mark.parametrize("extrapolation_type", [None, "linear", "constant"])
+def test_not_enough(count, extrapolation_type):
+    # This also tests that the single value and constant extrapolation edge-case
+    # doesn't work if nan's are involved
     source = np.asarray(
         [
             np.datetime64("1000-01-01"),
@@ -119,7 +129,11 @@ def test_not_enough():
         ],
         dtype=np.datetime64,
     )
-    source_vals = [1.0, 2.0, np.nan, np.nan]
+    source_vals = [np.nan] * 4
+    # Set the first 'count' values to be valid
+    for i in range(count):
+        source_vals[i] = i
+
     target = np.asarray(
         [
             np.datetime64("1000-01-01"),
@@ -129,9 +143,60 @@ def test_not_enough():
         ],
         dtype=np.datetime64,
     )
-    c = TimeseriesConverter(
-        source,
-        target,
+    c = TimeseriesConverter(source, target, extrapolation_type=extrapolation_type)
+    with pytest.raises(InsufficientDataError):
+        c.convert_from(source_vals)
+
+
+def test_constant_extrapolation():
+    source = np.asarray(
+        [
+            np.datetime64("2000-01-01"),
+        ],
+        dtype=np.datetime64,
     )
+    source_vals = [1.0]
+    target = np.asarray(
+        [
+            np.datetime64("1000-01-01"),
+            np.datetime64("2000-01-01"),
+            np.datetime64("3000-01-01"),
+        ],
+        dtype=np.datetime64,
+    )
+
+    # Linear extrapolation (default) should fail
+    c = TimeseriesConverter(source, target)
+    with pytest.raises(InsufficientDataError):
+        c.convert_from(source_vals)
+    c = TimeseriesConverter(source, target, extrapolation_type="linear")
+    with pytest.raises(InsufficientDataError):
+        c.convert_from(source_vals)
+
+    # Constant extrapolation should work with a single value
+    c = TimeseriesConverter(source, target, extrapolation_type="constant")
+    res = c.convert_from(source_vals)
+
+    np.testing.assert_allclose(res, [1.0, 1.0, 1.0])
+
+
+def test_constant_extrapolation_nan():
+    source = np.asarray(
+        [
+            np.datetime64("2000-01-01"),
+        ],
+        dtype=np.datetime64,
+    )
+    source_vals = [np.nan]
+    target = np.asarray(
+        [
+            np.datetime64("1000-01-01"),
+            np.datetime64("2000-01-01"),
+            np.datetime64("3000-01-01"),
+        ],
+        dtype=np.datetime64,
+    )
+
+    c = TimeseriesConverter(source, target, extrapolation_type="constant")
     with pytest.raises(InsufficientDataError):
         c.convert_from(source_vals)